Update feed component keys
The `EntryCluster` `ul li` component key hash resulted in collisions
when, for example, multiple clusters would appear with the same season,
day, and phase. This can occur when a day switches between phases, such
as going from `1` to `3` and back to `1`. Adding the event ID of the
first event in the cluster fixes this component key collision.

This commit also cleans up a few existing components which feature
either superfluous keys or missing keys.